Changes to the trajectory generation system. Although the idea of implementing functions as the building block of trajectories remains and much of the higher level interface is similar, the philosophy between the functions and the data has changed.
First, robSpace is used to define the space of each variable and robMapping is
used to define the domain and codomain of each function. These classes are used
to ensure consistency between the functions and the variables. Second, the 
representation of trajectories is more constrained. In this version, every 
trajectory is broken down to a 1D function or a function over SO(3). Although
this can create redundancy in evaluating the function it reduces the number of
classes that are required for a given trajectory.

git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/main@1246 0959eb1a-6956-4f60-947b-6092a4e22c97