Call AdvanceIfAutomatic after Startup too
In StartupInternal, just prior to the RunEvent() call, update the state
tables that have automatic advance turned on, just like the similar
code at the bottom of RunInternal.

Allows components to modify their state data in the Startup call and
have it immediately available for reading (prior to the completion of
the first Run() call) without requiring manual Start/Advance calls.

If the cisst framework advances the state tables automatically after
Run(), it should also do it after Startup() for consistency.