Fix plot scale for (near) constant plots
For plot signals whose data Y-range is small enough to be effectively constant (depends on plot view port size, but roughly less than 1e-300), the computed scale for the Y axis was overflowing to +infinity. An infinite scale then results in the plot not being rendered, so this adds checks to set the X/Y scales to 1.0 if they are infinite.